rename to query_planning
