+++ mypy.ini
@@ -160,6 +160,9 @@ disallow_untyped_calls = False
 [mypy-graphql_compiler.query_pagination.query_parameterizer.*]
 disallow_untyped_calls = False
 
+[mypy-graphql_compiler.query_planning.*]
+disallow_untyped_defs = False
+
 [mypy-graphql_compiler.schema_generation.graphql_schema.*]
 check_untyped_defs = False
 disallow_incomplete_defs = False
@@ -188,9 +191,6 @@ check_untyped_defs = False
 
 [mypy-graphql_compiler.schema_transformation.*]
 disallow_untyped_calls = False
-
-[mypy-graphql_compiler.schema_transformation.make_query_plan.*]
-disallow_untyped_defs = False
 
 [mypy-graphql_compiler.schema_transformation.split_query.*]
 disallow_incomplete_defs = False
